+++ core/full_coverage_test.go
@@ -22,7 +22,7 @@ func TestDownloadBundleFull(t *testing.T) {
 	defer func() { allPatterns = originalPatterns }()
 
 	// Test actual download (may fail in test environment)
-	err := DownloadBundle(context.Background())
+	err := DownloadBundle(context.Background(), false)
 	if err != nil {
 		t.Logf("DownloadBundle failed (expected in test environment): %v", err)
 		// This is expected - we can't easily mock HTTP without refactoring